Refactor and move AABB to utils/geometry to accommodate new colliders
