+++ library/cpp/include/wavemap/core/indexing/index_conversions.h
@@ -6,10 +6,10 @@
 #include <numeric>
 
 #include "wavemap/core/common.h"
-#include "wavemap/core/data_structure/aabb.h"
 #include "wavemap/core/indexing/ndtree_index.h"
 #include "wavemap/core/utils/bits/morton_encoding.h"
 #include "wavemap/core/utils/data/eigen_checks.h"
+#include "wavemap/core/utils/geometry/aabb.h"
 #include "wavemap/core/utils/math/int_math.h"
 
 namespace wavemap::convert {
