+++ library/cpp/include/wavemap/core/integrator/projective/coarse_to_fine/range_image_intersector.h
@@ -6,12 +6,12 @@
 #include <utility>
 
 #include "wavemap/core/common.h"
-#include "wavemap/core/data_structure/aabb.h"
 #include "wavemap/core/data_structure/image.h"
 #include "wavemap/core/integrator/measurement_model/measurement_model_base.h"
 #include "wavemap/core/integrator/projection_model/projector_base.h"
 #include "wavemap/core/integrator/projective/coarse_to_fine/hierarchical_range_bounds.h"
 #include "wavemap/core/integrator/projective/update_type.h"
+#include "wavemap/core/utils/geometry/aabb.h"
 
 namespace wavemap {
 class RangeImageIntersector {
